+++ src/adapters/google.js
@@ -129,7 +129,7 @@ export default class {
       let options = jsOptions(chart, chart.options, chartOptions);
       let data = this.createDataTable(chart.data, chart.xtype);
 
-      this.drawChart(chart, this.library.visualization.LineChart, data, options);
+      this.drawChart(chart, "LineChart", data, options);
     });
   }
 
@@ -161,7 +161,7 @@ export default class {
       data.addColumn("number", "Value");
       data.addRows(chart.data);
 
-      this.drawChart(chart, this.library.visualization.PieChart, data, options);
+      this.drawChart(chart, "PieChart", data, options);
     });
   }
 
@@ -170,7 +170,7 @@ export default class {
       let options = jsOptions(chart, chart.options);
       let data = this.createDataTable(chart.data, chart.xtype);
 
-      this.drawChart(chart, this.library.visualization.ColumnChart, data, options);
+      this.drawChart(chart, "ColumnChart", data, options);
     });
   }
 
@@ -186,7 +186,7 @@ export default class {
       let options = jsOptionsFunc(defaultOptions, hideLegend, setTitle, setBarMin, setBarMax, setStacked, setXtitle, setYtitle)(chart, chart.options, chartOptions);
       let data = this.createDataTable(chart.data, chart.xtype);
 
-      this.drawChart(chart, this.library.visualization.BarChart, data, options);
+      this.drawChart(chart, "BarChart", data, options);
     });
   }
 
@@ -201,7 +201,7 @@ export default class {
       let options = jsOptions(chart, chart.options, chartOptions);
       let data = this.createDataTable(chart.data, chart.xtype);
 
-      this.drawChart(chart, this.library.visualization.AreaChart, data, options);
+      this.drawChart(chart, "AreaChart", data, options);
     });
   }
 
@@ -220,7 +220,7 @@ export default class {
       data.addColumn("number", chart.options.label || "Value");
       data.addRows(chart.data);
 
-      this.drawChart(chart, this.library.visualization.GeoChart, data, options);
+      this.drawChart(chart, "GeoChart", data, options);
     });
   }
 
@@ -248,7 +248,7 @@ export default class {
       }
       data.addRows(rows2);
 
-      this.drawChart(chart, this.library.visualization.ScatterChart, data, options);
+      this.drawChart(chart, "ScatterChart", data, options);
     });
   }
 
@@ -271,19 +271,27 @@ export default class {
 
       chart.element.style.lineHeight = "normal";
 
-      this.drawChart(chart, this.library.visualization.Timeline, data, options);
+      this.drawChart(chart, "Timeline", data, options);
     });
   }
 
-  drawChart(chart, type, data, options) {
+  destroy(chart) {
     if (chart.chart) {
       chart.chart.clearChart();
     }
+  }
 
-    chart.chart = new type(chart.element);
-    resize(function () {
-      chart.chart.draw(data, options);
-    });
+  drawChart(chart, type, data, options) {
+    this.destroy(chart);
+
+    if (chart.options.eject) {
+      chart.element.innerText = "var data = new google.visualization.DataTable(" + data.toJSON() + ");\nvar chart = new google.visualization." + type + "(element);\nchart.draw(data, " + JSON.stringify(options) + ");";
+    } else {
+      chart.chart = new this.library.visualization[type](chart.element);
+      resize(function () {
+        chart.chart.draw(data, options);
+      });
+    }
   }
 
   waitForLoaded(chart, pack, callback) {